Moving files around for better modularization
