+++ benchmarks/pdx_insertion.cpp
@@ -9,7 +9,8 @@
 #include <vector>
 
 #include "benchmark_utils.hpp"
-#include "pdx/index.hpp"
+#include "pdx/indexes/ivf_tree.hpp"
+#include "pdx/indexes/ivf_vanilla.hpp"
 #include "pdx/profiler.hpp"
 #include "pdx/utils.hpp"
