+++ benchmarks/pdx_ivf.cpp
@@ -1,5 +1,6 @@
 #include "benchmark_utils.hpp"
-#include "pdx/index.hpp"
+#include "pdx/indexes/ivf_tree.hpp"
+#include "pdx/indexes/ivf_vanilla.hpp"
 #include "pdx/utils.hpp"
 #include <iomanip>
 #include <iostream>
