+++ benchmarks/pdx_serialization.cpp
@@ -7,7 +7,8 @@
 #include <vector>
 
 #include "benchmark_utils.hpp"
-#include "pdx/index.hpp"
+#include "pdx/indexes/ivf_tree.hpp"
+#include "pdx/indexes/ivf_vanilla.hpp"
 
 template <typename IndexT>
 void BuildAndSave(
