+++ include/pdx/indexes/ivf_tree.hpp
@@ -13,8 +13,9 @@
 
 #include "pdx/clustering.hpp"
 #include "pdx/common.hpp"
-#include "pdx/ivf_wrapper.hpp"
-#include "pdx/layout.hpp"
+#include "pdx/indexes/ivf_core.hpp"
+#include "pdx/indexes/ivf_utils.hpp"
+#include "pdx/indexes/ivf_vanilla.hpp"
 #include "pdx/profiler.hpp"
 #include "pdx/pruners/adsampling.hpp"
 #include "pdx/quantizers/scalar.hpp"
@@ -23,428 +24,6 @@
 #include <omp.h>
 
 namespace PDX {
-
-struct PDXIndexConfig {
-    uint32_t num_dimensions;
-    DistanceMetric distance_metric = DistanceMetric::L2SQ;
-    uint32_t seed = 42;
-    uint32_t num_clusters = 0; // 0 = auto-compute from num_embeddings
-    uint32_t num_meso_clusters = 0;
-    bool normalize = false;
-    float sampling_fraction = 0.0f; // 0 = auto (1.0 if small dataset, 0.3 otherwise)
-    uint32_t kmeans_iters = 10;
-    bool hierarchical_indexing = true;
-    uint32_t n_threads = 0; // 0 = omp_get_max_threads()
-
-    void Validate() const {
-        if (num_dimensions == 0 || num_dimensions > PDX_MAX_DIMS) {
-            throw std::invalid_argument(
-                "num_dimensions must be between 1 and " + std::to_string(PDX_MAX_DIMS) + ", got " +
-                std::to_string(num_dimensions)
-            );
-        }
-        if (sampling_fraction < 0.0f || sampling_fraction > 1.0f) {
-            throw std::invalid_argument(
-                "sampling_fraction must be between 0.0 and 1.0, got " +
-                std::to_string(sampling_fraction)
-            );
-        }
-        if (num_meso_clusters > 0 && num_clusters > 0 && num_meso_clusters >= num_clusters) {
-            throw std::invalid_argument(
-                "num_meso_clusters (" + std::to_string(num_meso_clusters) +
-                ") must be smaller than num_clusters (" + std::to_string(num_clusters) + ")"
-            );
-        }
-        if (kmeans_iters == 0 || kmeans_iters >= 100) {
-            throw std::invalid_argument(
-                "kmeans_iters must be between 1 and 99, got " + std::to_string(kmeans_iters)
-            );
-        }
-    }
-
-    void ValidateNumEmbeddings(size_t num_embeddings) const {
-        if (num_clusters > 0 && num_clusters > num_embeddings) {
-            throw std::invalid_argument(
-                "num_clusters (" + std::to_string(num_clusters) + ") exceeds num_embeddings (" +
-                std::to_string(num_embeddings) + ")"
-            );
-        }
-    }
-};
-
-inline std::unique_ptr<float[]> NormalizeAndRotate(
-    const float* embeddings,
-    size_t num_embeddings,
-    uint32_t num_dimensions,
-    bool normalize,
-    const ADSamplingPruner& pruner
-) {
-    const size_t total_floats = num_embeddings * num_dimensions;
-    std::unique_ptr<float[]> normalized;
-    const float* rotation_input = embeddings;
-    if (normalize) {
-        normalized.reset(new float[total_floats]);
-        Quantizer quantizer(num_dimensions);
-#pragma omp parallel for if (num_embeddings > 1) num_threads(PDX::g_n_threads)
-        for (size_t i = 0; i < num_embeddings; i++) {
-            quantizer.NormalizeQuery(
-                embeddings + i * num_dimensions, normalized.get() + i * num_dimensions
-            );
-        }
-        rotation_input = normalized.get();
-    }
-    std::unique_ptr<float[]> preprocessed(new float[total_floats]);
-    pruner.PreprocessEmbeddings(rotation_input, preprocessed.get(), num_embeddings);
-    return preprocessed;
-}
-
-template <Quantization Q>
-void PopulateIVFClusters(
-    IVF<Q>& ivf,
-    const KMeansResult& kmeans_result,
-    const float* source_data,
-    const size_t* row_ids,
-    uint32_t num_dimensions,
-    uint32_t num_clusters,
-    float quantization_base,
-    float quantization_scale
-) {
-    using storage_t = pdx_data_t<Q>;
-
-    size_t max_cluster_size = 0;
-    for (size_t i = 0; i < num_clusters; i++) {
-        max_cluster_size = std::max(max_cluster_size, kmeans_result.assignments[i].size());
-    }
-
-    // Pre-allocate all clusters sequentially
-    for (size_t cluster_idx = 0; cluster_idx < num_clusters; cluster_idx++) {
-        ivf.clusters.emplace_back(kmeans_result.assignments[cluster_idx].size(), num_dimensions);
-        ivf.clusters[cluster_idx].id = cluster_idx;
-    }
-
-    // Per-thread tmp buffers for gather + quantize
-    const uint32_t n_threads = PDX::g_n_threads;
-    std::vector<std::unique_ptr<storage_t[]>> tmp_buffers(n_threads);
-    for (uint32_t t = 0; t < n_threads; t++) {
-        tmp_buffers[t].reset(new storage_t[static_cast<uint64_t>(max_cluster_size) * num_dimensions]
-        );
-    }
-
-#pragma omp parallel for num_threads(n_threads)
-    for (size_t cluster_idx = 0; cluster_idx < num_clusters; cluster_idx++) {
-        const auto cluster_size = kmeans_result.assignments[cluster_idx].size();
-        auto& cluster = ivf.clusters[cluster_idx];
-        auto* tmp = tmp_buffers[omp_get_thread_num()].get();
-
-        for (size_t pos = 0; pos < cluster_size; pos++) {
-            const auto emb_idx = kmeans_result.assignments[cluster_idx][pos];
-            cluster.indices[pos] = row_ids[emb_idx];
-
-            if constexpr (Q == U8) {
-                ScalarQuantizer<Q> quantizer(num_dimensions);
-                quantizer.QuantizeEmbedding(
-                    source_data + (emb_idx * num_dimensions),
-                    quantization_base,
-                    quantization_scale,
-                    tmp + (pos * num_dimensions)
-                );
-            } else {
-                std::memcpy(
-                    tmp + (pos * num_dimensions),
-                    source_data + (emb_idx * num_dimensions),
-                    num_dimensions * sizeof(float)
-                );
-            }
-        }
-        StoreClusterEmbeddings<Q, storage_t>(cluster, ivf, tmp, cluster_size);
-    }
-
-    ivf.ComputeClusterOffsets();
-}
-
-class IPDXIndex {
-  public:
-    virtual ~IPDXIndex() = default;
-    virtual std::vector<KNNCandidate> Search(const float* query_embedding, size_t knn) const = 0;
-    virtual std::vector<KNNCandidate> FilteredSearch(
-        const float* query_embedding,
-        size_t knn,
-        const std::vector<size_t>& passing_row_ids
-    ) const = 0;
-    virtual void BuildIndex(const float* embeddings, size_t num_embeddings) = 0;
-    virtual void SetNProbe(uint32_t n_probe) const = 0;
-    virtual void Save(const std::string& path) = 0;
-    virtual void Restore(const std::string& path) = 0;
-    virtual uint32_t GetNumDimensions() const = 0;
-    virtual uint32_t GetNumClusters() const = 0;
-    virtual uint32_t GetClusterSize(uint32_t cluster_id) const = 0;
-    virtual std::vector<uint32_t> GetClusterRowIds(uint32_t cluster_id) const = 0;
-    virtual size_t GetInMemorySizeInBytes() const = 0;
-    virtual void Append(size_t /*row_id*/, const float* /*embedding*/) {
-        throw std::runtime_error("Append is not supported by this index type. Use PDXTreeIndex.");
-    }
-    virtual void Delete(size_t /*row_id*/) {
-        throw std::runtime_error("Delete is not supported by this index type. Use PDXTreeIndex.");
-    }
-};
-
-template <PDX::Quantization Q>
-class PDXIndex : public IPDXIndex {
-  public:
-    using embedding_storage_t = PDX::pdx_data_t<Q>;
-    using cluster_t = PDX::Cluster<Q>;
-
-  private:
-    PDXIndexConfig config{};
-    PDX::IVF<Q> index;
-    std::unique_ptr<PDX::ADSamplingPruner> pruner;
-    std::unique_ptr<PDX::PDXearch<Q>> searcher;
-    std::vector<std::pair<uint32_t, uint32_t>> row_id_cluster_mapping;
-
-  public:
-    PDXIndex() = default;
-
-    explicit PDXIndex(PDXIndexConfig config) : config(config) {
-        config.Validate();
-        PDX::g_n_threads = (config.n_threads == 0) ? omp_get_max_threads() : config.n_threads;
-        pruner = std::make_unique<PDX::ADSamplingPruner>(config.num_dimensions, config.seed);
-    }
-
-    void Save(const std::string& path) override {
-        // Compact all clusters before saving
-        for (uint32_t c = 0; c < index.num_clusters; c++) {
-            auto moves = index.clusters[c].CompactCluster();
-            for (const auto& [row_id, new_idx] : moves) {
-                row_id_cluster_mapping[row_id] = {c, new_idx};
-            }
-        }
-
-        std::ofstream out(path, std::ios::binary);
-
-        uint8_t type_flag = static_cast<uint8_t>(GetIndexType());
-        out.write(reinterpret_cast<const char*>(&type_flag), sizeof(uint8_t));
-
-        // Rotation matrix
-        const auto& matrix = pruner->GetMatrix();
-        uint32_t matrix_rows = static_cast<uint32_t>(matrix.rows());
-        uint32_t matrix_cols = static_cast<uint32_t>(matrix.cols());
-        out.write(reinterpret_cast<const char*>(&matrix_rows), sizeof(uint32_t));
-        out.write(reinterpret_cast<const char*>(&matrix_cols), sizeof(uint32_t));
-        out.write(
-            reinterpret_cast<const char*>(matrix.data()), sizeof(float) * matrix_rows * matrix_cols
-        );
-
-        // IVF data
-        index.Save(out);
-    }
-
-    void Restore(const std::string& path) override {
-        auto buffer = MmapFile(path);
-        char* ptr = buffer.get();
-
-        // Index type flag
-        ptr += sizeof(uint8_t);
-
-        // Rotation matrix (ptr may be misaligned after the uint8_t type flag)
-        uint32_t matrix_rows, matrix_cols;
-        std::memcpy(&matrix_rows, ptr, sizeof(uint32_t));
-        ptr += sizeof(uint32_t);
-        std::memcpy(&matrix_cols, ptr, sizeof(uint32_t));
-        ptr += sizeof(uint32_t);
-        const size_t matrix_floats = static_cast<size_t>(matrix_rows) * matrix_cols;
-        auto aligned_matrix = std::unique_ptr<float[]>(new float[matrix_floats]);
-        std::memcpy(aligned_matrix.get(), ptr, sizeof(float) * matrix_floats);
-        ptr += sizeof(float) * matrix_floats;
-
-        // Load IVF data
-        index.Load(ptr);
-
-        // Create pruner and searcher
-        pruner =
-            std::make_unique<PDX::ADSamplingPruner>(index.num_dimensions, aligned_matrix.get());
-        searcher = std::make_unique<PDX::PDXearch<Q>>(index, *pruner);
-        BuildRowIdClusterMapping();
-    }
-
-    std::vector<PDX::KNNCandidate> Search(const float* query_embedding, size_t knn) const override {
-        return searcher->Search(query_embedding, knn);
-    }
-
-    std::vector<PDX::KNNCandidate> FilteredSearch(
-        const float* query_embedding,
-        size_t knn,
-        const std::vector<size_t>& passing_row_ids
-    ) const override {
-        auto evaluator = CreatePredicateEvaluator(passing_row_ids);
-        return searcher->FilteredSearch(query_embedding, knn, evaluator);
-    }
-
-    void SetNProbe(uint32_t n_probe) const override { searcher->SetNProbe(n_probe); }
-
-    const PDX::PDXearch<Q>& GetSearcher() const { return *searcher; }
-
-    uint32_t GetNumDimensions() const override { return index.num_dimensions; }
-
-    uint32_t GetNumClusters() const override { return index.num_clusters; }
-
-    uint32_t GetClusterSize(uint32_t cluster_id) const override {
-        return index.clusters[cluster_id].num_embeddings;
-    }
-
-    std::vector<uint32_t> GetClusterRowIds(uint32_t cluster_id) const override {
-        const auto& cluster = index.clusters[cluster_id];
-        std::vector<uint32_t> row_ids;
-        row_ids.reserve(cluster.num_embeddings);
-        for (uint32_t i = 0; i < cluster.used_capacity; i++) {
-            if (!cluster.HasTombstone(i)) {
-                row_ids.push_back(cluster.indices[i]);
-            }
-        }
-        return row_ids;
-    }
-
-    size_t GetInMemorySizeInBytes() const override {
-        size_t size = sizeof(*this);
-        // IVF heap allocations (sizeof(IVF<Q>) is inline in sizeof(*this))
-        size += index.GetInMemorySizeInBytes() - sizeof(index);
-        // Pruner: rotation matrix or flip_masks (DCT mode) + ratios vector
-        if (pruner) {
-            size += sizeof(*pruner);
-            const auto& m = pruner->GetMatrix();
-            // matrix heap data (1 x D for DCT sign vector, D x D for full rotation)
-            size += static_cast<size_t>(m.rows()) * m.cols() * sizeof(float);
-            size += pruner->num_dimensions * sizeof(float); // ratios
-            if (m.rows() == 1) {
-                size += pruner->num_dimensions * sizeof(uint32_t); // flip_masks
-            }
-        }
-        if (searcher) {
-            size += sizeof(*searcher);
-        }
-        // Row ID to cluster mapping
-        size += row_id_cluster_mapping.capacity() * sizeof(std::pair<uint32_t, uint32_t>);
-        return size;
-    }
-
-    void BuildIndex(const float* const embeddings, const size_t num_embeddings) override {
-        std::vector<size_t> row_ids(num_embeddings);
-        std::iota(row_ids.begin(), row_ids.end(), 0);
-        BuildIndex(row_ids.data(), embeddings, num_embeddings);
-    }
-
-    void BuildIndex(
-        const size_t* const row_ids,
-        const float* const embeddings,
-        const size_t num_embeddings
-    ) {
-        config.ValidateNumEmbeddings(num_embeddings);
-
-        const auto num_dimensions = config.num_dimensions;
-        auto num_clusters = config.num_clusters;
-        if (num_clusters == 0) {
-            num_clusters = ComputeNumberOfClusters(num_embeddings);
-        }
-        const bool normalize =
-            config.normalize || DistanceMetricRequiresNormalization(config.distance_metric);
-
-        assert(num_embeddings > 0);
-        assert(pruner);
-
-        auto preprocessed =
-            NormalizeAndRotate(embeddings, num_embeddings, num_dimensions, normalize, *pruner);
-
-        float quantization_base = 0.0f;
-        float quantization_scale = 1.0f;
-        if constexpr (Q == PDX::U8) {
-            const auto params = PDX::ScalarQuantizer<Q>::ComputeQuantizationParams(
-                preprocessed.get(), static_cast<size_t>(num_embeddings) * num_dimensions
-            );
-            quantization_base = params.quantization_base;
-            quantization_scale = params.quantization_scale;
-            index = PDX::IVF<Q>(
-                num_dimensions,
-                num_embeddings,
-                num_clusters,
-                normalize,
-                quantization_scale,
-                quantization_base
-            );
-        } else {
-            index = PDX::IVF<Q>(num_dimensions, num_embeddings, num_clusters, normalize);
-        }
-
-        KMeansResult kmeans_result = ComputeKMeans(
-            preprocessed.get(),
-            num_embeddings,
-            num_dimensions,
-            num_clusters,
-            config.distance_metric,
-            config.seed,
-            config.normalize,
-            config.sampling_fraction,
-            config.kmeans_iters,
-            config.hierarchical_indexing
-        );
-        index.centroids = std::move(kmeans_result.centroids);
-
-        PopulateIVFClusters<Q>(
-            index,
-            kmeans_result,
-            preprocessed.get(),
-            row_ids,
-            num_dimensions,
-            num_clusters,
-            quantization_base,
-            quantization_scale
-        );
-
-        searcher = std::make_unique<PDX::PDXearch<Q>>(index, *pruner);
-        BuildRowIdClusterMapping();
-    }
-
-    void Append(size_t /*row_id*/, const float* /*embedding*/) override {
-        throw std::runtime_error("Append is not implemented in PDXIndex. Use PDXTreeIndex instead."
-        );
-    }
-
-    void Delete(size_t /*row_id*/) override {
-        throw std::runtime_error("Delete is not implemented in PDXIndex. Use PDXTreeIndex instead."
-        );
-    }
-
-  private:
-    static constexpr PDXIndexType GetIndexType() {
-        if constexpr (Q == F32)
-            return PDXIndexType::PDX_F32;
-        else
-            return PDXIndexType::PDX_U8;
-    }
-
-    void BuildRowIdClusterMapping() {
-        size_t total = 0;
-        for (size_t c = 0; c < index.num_clusters; c++) {
-            total += index.clusters[c].num_embeddings;
-        }
-        row_id_cluster_mapping.resize(total);
-        for (uint32_t c = 0; c < index.num_clusters; c++) {
-            for (uint32_t p = 0; p < index.clusters[c].num_embeddings; p++) {
-                row_id_cluster_mapping[index.clusters[c].indices[p]] = {c, p};
-            }
-        }
-    }
-
-    PDX::PredicateEvaluator CreatePredicateEvaluator(const std::vector<size_t>& passing_row_ids
-    ) const {
-        PDX_PROFILE_SCOPE("PredicateEvaluator");
-        PDX::PredicateEvaluator evaluator(index.num_clusters, index.total_capacity);
-        for (const auto row_id : passing_row_ids) {
-            const auto& [cluster_id, index_in_cluster] = row_id_cluster_mapping[row_id];
-            evaluator.n_passing_tuples[cluster_id]++;
-            evaluator.selection_vector[index.cluster_offsets[cluster_id] + index_in_cluster] = 1;
-        }
-        return evaluator;
-    }
-};
 
 template <PDX::Quantization Q>
 class PDXTreeIndex : public IPDXIndex {
