+++ include/pdx/indexes/ivf_vanilla.hpp
@@ -1,0 +1,306 @@
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+#include <cstring>
+#include <fstream>
+#include <limits>
+#include <memory>
+#include <numeric>
+#include <stdexcept>
+#include <string>
+#include <utility>
+
+#include "pdx/common.hpp"
+#include "pdx/indexes/ivf_utils.hpp"
+#include "pdx/pruners/adsampling.hpp"
+#include "pdx/quantizers/scalar.hpp"
+#include "pdx/searcher.hpp"
+
+namespace PDX {
+
+class IPDXIndex {
+  public:
+    virtual ~IPDXIndex() = default;
+    virtual std::vector<KNNCandidate> Search(const float* query_embedding, size_t knn) const = 0;
+    virtual std::vector<KNNCandidate> FilteredSearch(
+        const float* query_embedding,
+        size_t knn,
+        const std::vector<size_t>& passing_row_ids
+    ) const = 0;
+    virtual void BuildIndex(const float* embeddings, size_t num_embeddings) = 0;
+    virtual void SetNProbe(uint32_t n_probe) const = 0;
+    virtual void Save(const std::string& path) = 0;
+    virtual void Restore(const std::string& path) = 0;
+    virtual uint32_t GetNumDimensions() const = 0;
+    virtual uint32_t GetNumClusters() const = 0;
+    virtual uint32_t GetClusterSize(uint32_t cluster_id) const = 0;
+    virtual std::vector<uint32_t> GetClusterRowIds(uint32_t cluster_id) const = 0;
+    virtual size_t GetInMemorySizeInBytes() const = 0;
+    virtual void Append(size_t /*row_id*/, const float* /*embedding*/) {
+        throw std::runtime_error("Append is not supported by this index type. Use PDXTreeIndex.");
+    }
+    virtual void Delete(size_t /*row_id*/) {
+        throw std::runtime_error("Delete is not supported by this index type. Use PDXTreeIndex.");
+    }
+};
+
+template <PDX::Quantization Q>
+class PDXIndex : public IPDXIndex {
+  public:
+    using embedding_storage_t = PDX::pdx_data_t<Q>;
+    using cluster_t = PDX::Cluster<Q>;
+
+  private:
+    PDXIndexConfig config{};
+    PDX::IVF<Q> index;
+    std::unique_ptr<PDX::ADSamplingPruner> pruner;
+    std::unique_ptr<PDX::PDXearch<Q>> searcher;
+    std::vector<std::pair<uint32_t, uint32_t>> row_id_cluster_mapping;
+
+  public:
+    PDXIndex() = default;
+
+    explicit PDXIndex(PDXIndexConfig config) : config(config) {
+        config.Validate();
+        PDX::g_n_threads = (config.n_threads == 0) ? omp_get_max_threads() : config.n_threads;
+        pruner = std::make_unique<PDX::ADSamplingPruner>(config.num_dimensions, config.seed);
+    }
+
+    void Save(const std::string& path) override {
+        // Compact all clusters before saving
+        for (uint32_t c = 0; c < index.num_clusters; c++) {
+            auto moves = index.clusters[c].CompactCluster();
+            for (const auto& [row_id, new_idx] : moves) {
+                row_id_cluster_mapping[row_id] = {c, new_idx};
+            }
+        }
+
+        std::ofstream out(path, std::ios::binary);
+
+        uint8_t type_flag = static_cast<uint8_t>(GetIndexType());
+        out.write(reinterpret_cast<const char*>(&type_flag), sizeof(uint8_t));
+
+        // Rotation matrix
+        const auto& matrix = pruner->GetMatrix();
+        uint32_t matrix_rows = static_cast<uint32_t>(matrix.rows());
+        uint32_t matrix_cols = static_cast<uint32_t>(matrix.cols());
+        out.write(reinterpret_cast<const char*>(&matrix_rows), sizeof(uint32_t));
+        out.write(reinterpret_cast<const char*>(&matrix_cols), sizeof(uint32_t));
+        out.write(
+            reinterpret_cast<const char*>(matrix.data()), sizeof(float) * matrix_rows * matrix_cols
+        );
+
+        // IVF data
+        index.Save(out);
+    }
+
+    void Restore(const std::string& path) override {
+        auto buffer = MmapFile(path);
+        char* ptr = buffer.get();
+
+        // Index type flag
+        ptr += sizeof(uint8_t);
+
+        // Rotation matrix (ptr may be misaligned after the uint8_t type flag)
+        uint32_t matrix_rows, matrix_cols;
+        std::memcpy(&matrix_rows, ptr, sizeof(uint32_t));
+        ptr += sizeof(uint32_t);
+        std::memcpy(&matrix_cols, ptr, sizeof(uint32_t));
+        ptr += sizeof(uint32_t);
+        const size_t matrix_floats = static_cast<size_t>(matrix_rows) * matrix_cols;
+        auto aligned_matrix = std::unique_ptr<float[]>(new float[matrix_floats]);
+        std::memcpy(aligned_matrix.get(), ptr, sizeof(float) * matrix_floats);
+        ptr += sizeof(float) * matrix_floats;
+
+        // Load IVF data
+        index.Load(ptr);
+
+        // Create pruner and searcher
+        pruner =
+            std::make_unique<PDX::ADSamplingPruner>(index.num_dimensions, aligned_matrix.get());
+        searcher = std::make_unique<PDX::PDXearch<Q>>(index, *pruner);
+        BuildRowIdClusterMapping();
+    }
+
+    std::vector<PDX::KNNCandidate> Search(const float* query_embedding, size_t knn) const override {
+        return searcher->Search(query_embedding, knn);
+    }
+
+    std::vector<PDX::KNNCandidate> FilteredSearch(
+        const float* query_embedding,
+        size_t knn,
+        const std::vector<size_t>& passing_row_ids
+    ) const override {
+        auto evaluator = CreatePredicateEvaluator(passing_row_ids);
+        return searcher->FilteredSearch(query_embedding, knn, evaluator);
+    }
+
+    void SetNProbe(uint32_t n_probe) const override { searcher->SetNProbe(n_probe); }
+
+    const PDX::PDXearch<Q>& GetSearcher() const { return *searcher; }
+
+    uint32_t GetNumDimensions() const override { return index.num_dimensions; }
+
+    uint32_t GetNumClusters() const override { return index.num_clusters; }
+
+    uint32_t GetClusterSize(uint32_t cluster_id) const override {
+        return index.clusters[cluster_id].num_embeddings;
+    }
+
+    std::vector<uint32_t> GetClusterRowIds(uint32_t cluster_id) const override {
+        const auto& cluster = index.clusters[cluster_id];
+        std::vector<uint32_t> row_ids;
+        row_ids.reserve(cluster.num_embeddings);
+        for (uint32_t i = 0; i < cluster.used_capacity; i++) {
+            if (!cluster.HasTombstone(i)) {
+                row_ids.push_back(cluster.indices[i]);
+            }
+        }
+        return row_ids;
+    }
+
+    size_t GetInMemorySizeInBytes() const override {
+        size_t size = sizeof(*this);
+        // IVF heap allocations (sizeof(IVF<Q>) is inline in sizeof(*this))
+        size += index.GetInMemorySizeInBytes() - sizeof(index);
+        // Pruner: rotation matrix or flip_masks (DCT mode) + ratios vector
+        if (pruner) {
+            size += sizeof(*pruner);
+            const auto& m = pruner->GetMatrix();
+            // matrix heap data (1 x D for DCT sign vector, D x D for full rotation)
+            size += static_cast<size_t>(m.rows()) * m.cols() * sizeof(float);
+            size += pruner->num_dimensions * sizeof(float); // ratios
+            if (m.rows() == 1) {
+                size += pruner->num_dimensions * sizeof(uint32_t); // flip_masks
+            }
+        }
+        if (searcher) {
+            size += sizeof(*searcher);
+        }
+        // Row ID to cluster mapping
+        size += row_id_cluster_mapping.capacity() * sizeof(std::pair<uint32_t, uint32_t>);
+        return size;
+    }
+
+    void BuildIndex(const float* const embeddings, const size_t num_embeddings) override {
+        std::vector<size_t> row_ids(num_embeddings);
+        std::iota(row_ids.begin(), row_ids.end(), 0);
+        BuildIndex(row_ids.data(), embeddings, num_embeddings);
+    }
+
+    void BuildIndex(
+        const size_t* const row_ids,
+        const float* const embeddings,
+        const size_t num_embeddings
+    ) {
+        config.ValidateNumEmbeddings(num_embeddings);
+
+        const auto num_dimensions = config.num_dimensions;
+        auto num_clusters = config.num_clusters;
+        if (num_clusters == 0) {
+            num_clusters = ComputeNumberOfClusters(num_embeddings);
+        }
+        const bool normalize =
+            config.normalize || DistanceMetricRequiresNormalization(config.distance_metric);
+
+        assert(num_embeddings > 0);
+        assert(pruner);
+
+        auto preprocessed =
+            NormalizeAndRotate(embeddings, num_embeddings, num_dimensions, normalize, *pruner);
+
+        float quantization_base = 0.0f;
+        float quantization_scale = 1.0f;
+        if constexpr (Q == PDX::U8) {
+            const auto params = PDX::ScalarQuantizer<Q>::ComputeQuantizationParams(
+                preprocessed.get(), static_cast<size_t>(num_embeddings) * num_dimensions
+            );
+            quantization_base = params.quantization_base;
+            quantization_scale = params.quantization_scale;
+            index = PDX::IVF<Q>(
+                num_dimensions,
+                num_embeddings,
+                num_clusters,
+                normalize,
+                quantization_scale,
+                quantization_base
+            );
+        } else {
+            index = PDX::IVF<Q>(num_dimensions, num_embeddings, num_clusters, normalize);
+        }
+
+        KMeansResult kmeans_result = ComputeKMeans(
+            preprocessed.get(),
+            num_embeddings,
+            num_dimensions,
+            num_clusters,
+            config.distance_metric,
+            config.seed,
+            config.normalize,
+            config.sampling_fraction,
+            config.kmeans_iters,
+            config.hierarchical_indexing
+        );
+        index.centroids = std::move(kmeans_result.centroids);
+
+        PopulateIVFClusters<Q>(
+            index,
+            kmeans_result,
+            preprocessed.get(),
+            row_ids,
+            num_dimensions,
+            num_clusters,
+            quantization_base,
+            quantization_scale
+        );
+
+        searcher = std::make_unique<PDX::PDXearch<Q>>(index, *pruner);
+        BuildRowIdClusterMapping();
+    }
+
+    void Append(size_t /*row_id*/, const float* /*embedding*/) override {
+        throw std::runtime_error("Append is not implemented in PDXIndex. Use PDXTreeIndex instead."
+        );
+    }
+
+    void Delete(size_t /*row_id*/) override {
+        throw std::runtime_error("Delete is not implemented in PDXIndex. Use PDXTreeIndex instead."
+        );
+    }
+
+  private:
+    static constexpr PDXIndexType GetIndexType() {
+        if constexpr (Q == F32)
+            return PDXIndexType::PDX_F32;
+        else
+            return PDXIndexType::PDX_U8;
+    }
+
+    void BuildRowIdClusterMapping() {
+        size_t total = 0;
+        for (size_t c = 0; c < index.num_clusters; c++) {
+            total += index.clusters[c].num_embeddings;
+        }
+        row_id_cluster_mapping.resize(total);
+        for (uint32_t c = 0; c < index.num_clusters; c++) {
+            for (uint32_t p = 0; p < index.clusters[c].num_embeddings; p++) {
+                row_id_cluster_mapping[index.clusters[c].indices[p]] = {c, p};
+            }
+        }
+    }
+
+    PDX::PredicateEvaluator CreatePredicateEvaluator(const std::vector<size_t>& passing_row_ids
+    ) const {
+        PDX_PROFILE_SCOPE("PredicateEvaluator");
+        PDX::PredicateEvaluator evaluator(index.num_clusters, index.total_capacity);
+        for (const auto row_id : passing_row_ids) {
+            const auto& [cluster_id, index_in_cluster] = row_id_cluster_mapping[row_id];
+            evaluator.n_passing_tuples[cluster_id]++;
+            evaluator.selection_vector[index.cluster_offsets[cluster_id] + index_in_cluster] = 1;
+        }
+        return evaluator;
+    }
+};
+
+} // namespace PDX
