+++ tests/generate_test_ground_truth.cpp
@@ -4,7 +4,8 @@
 #include <iostream>
 #include <vector>
 
-#include "pdx/index.hpp"
+#include "pdx/indexes/ivf_tree.hpp"
+#include "pdx/indexes/ivf_vanilla.hpp"
 #include "test_utils.hpp"
 
 int main() {
